several changes - create `disabledSelectRow()` - move `rankArrayToSelectOptions()` to dAPIRequests - fix crashes in `/rank remove` - iterate on interaction acknowledgements in `/rank remove`
