+++ source/frontend/shared/messageParts.js
@@ -1,5 +1,5 @@
 const fs = require("fs");
-const { EmbedBuilder, Colors, Guild, ActionRowBuilder, ButtonBuilder, ButtonStyle, heading, userMention, MessageFlags, bold, italic, GuildMember, Role, Collection } = require("discord.js");
+const { EmbedBuilder, Colors, Guild, ActionRowBuilder, ButtonBuilder, ButtonStyle, heading, userMention, MessageFlags, bold, italic, GuildMember } = require("discord.js");
 const { MessageLimits, EmbedLimits } = require("@sapphire/discord.js-utilities");
 const { SAFE_DELIMITER } = require("../../constants");
 const { Bounty, Completion, Company, Season, Rank, Participation, Hunter } = require("../../database/models");
@@ -602,24 +602,6 @@ function generateSecondingRewardString(seconderDisplayName, recipientIds, rankUp
 	return text;
 }
 
-/**
- * @param {Rank[]} ranks
- * @param {Collection<string, Role>} allGuildRoles
- */
-function rankArrayToSelectOptions(ranks, allGuildRoles) {
-	return ranks.map((rank, index) => {
-		const option = {
-			label: rank.roleId ? allGuildRoles.get(rank.roleId).name : `Rank ${index + 1}`,
-			description: `Variance Threshold: ${rank.varianceThreshold}`,
-			value: rank.varianceThreshold.toString()
-		};
-		if (rank.rankmoji) {
-			option.emoji = rank.rankmoji;
-		}
-		return option;
-	})
-}
-
 module.exports = {
 	commandMention,
 	congratulationBuilder,
@@ -644,6 +626,5 @@ module.exports = {
 	generateSecondingActionRow,
 	generateToastRewardString,
 	generateCompletionEmbed,
-	generateSecondingRewardString,
-	rankArrayToSelectOptions
+	generateSecondingRewardString
 };
